Create video_0.sh
This is the video_0.sh program by Virtual Pilot Light. It is to be used with the video video_0.sh with the YouTube platform. Do not go to the video link! This is a BASH script. This is to be done on personal computers, as some content may be considered NSFW. Close all tabs on your browser. Git clone the repo on your own hard drive. Run the script. This is the only acceptable way to enjoy my media. Further questions should be aimed at the appropriate sources. I.E. do the research || don't speak.